Add more ResultTree tests and fix issues found
Add tests that ResultTree makes minimal system calls when element access
or slice filters are being used.

Fix two issues found in Filter.cpp by these tests.

Add a CMake option to force color output from the compiler and/or static
analyzer even when the output is not going to a terminal.