Remove ANTLR4 dependency
ANTRL4 isn't adding enough value to justify how big it is, how long it takes
to compile, or the JRE build-time dependency. Replace it with a
hand-written recursive descent parser that's only about as complex as
the code that used to convert the ANTLR4 parse tree into SQ AST.

It doesn't make much sense to have separate ast and parser libs now so
merge them.