Add a cache to system objects
In the near future there will be support for filtering lists of system
objects based on the values of their fields. In order for this to not
cause duplication of calls into the system, add a cache of field values
to system objects.

With the addition of the cache, it would become even messier to track a
unique owner for each system object, so just switch to using shared
ownership with std::shared_ptr. This simplifies a lot of of the
filtering code.

In the test code, tell GoogleMock to use better defaults for Results and
FieldPtrs so that we don't end up dereferencing a null pointer if a mock
object gets called unexpectedly.