Add tests for SqDataSize fields
Upon running the tests, the first thing to notice was that the error
message in the exception that SqDataSize's constructor throws if the
value is negative is terrible. Throw an OutOfRangeError rather than
relying on gsl::narrow to produce a good error message.

Make things easier by not converting between unsigned and signed types,
but instead just storing data sizes as PrimitiveInts (int64_ts) and
doing a "< 0" check in the constructor.