Start adding support for getting device info
Add support for getting some simple information about devices in the
system using libudev.

This functionality is Linux specific but there isn't a mechanism yet for
having different implementations for different platforms, or for
separating platform independent code from platform specific code... For
now, just rename "standard" to "linux" and always use that.

Update to a version of the build container that contains libudev.