Factor out some gsl::narrow error handling
There are a bunch of places that do
try {
    gsl::narrow<...>(...);
} catch(gsl::narrowing_error& e) {
    throw OutOfRangeError(...);
}
to provide a better error message when narrowing fails. Factor this out
into a gsl::narrow wrapper.

The natural place for the new narrow wrapper is util/typeutil.h, but it
needs to include errors.h for the type of exception to throw on
failure. I don't want the common_types and util libs to be mutually
dependent on each other and the line between "common_types" and "util"
is fairly arbitrary anyway so just combine both libraries into the new
"core" library.